requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0
faiss-cpu>=1.8.0
langchain>=0.2.10
boto3>=1.34.0
osmnx>=1.9.3
networkx>=3.3
streamlit>=1.37.0
requests>=2.32.0
